XS Fixes
The old version of the XS mangled the stack (I used helpers that were
not correct, oops). This takes the code I needed from the helpers, but
avoids POPSTACK so the stakc is not mangled.

I also added more tests to sysopen to make it easier to reproduce the
segv when I get it wrong. Turns out the mangling made it so that on my
system, the third call to open/sysopen is when the segv finally
happened. the cpantesters results that had segv were from systems where
the segv happened in 1 or 2 calls instead.